# verification/curator-service/api/openapi/openapi.yaml
openapi: 3.0.3
info:
    title: Global.health curator service API
    description: Server for CRUD operations on Global.health data
    termsOfService: 'https://global.health/terms-of-use/'
    contact:
        name: Global.health
        url: https://global.health
        email: info@global.health
    license:
        name: MIT
        url: 'https://opensource.org/licenses/MIT'
    version: 1.0.0
...
paths:
...
    /cases:
        get:
            tags: [Case]
            summary: Lists cases
            operationId: listCases
            parameters:
                - name: page
                  in: query
                  description: The pages of sources to skip before starting to collect the result set
                  required: false
                  schema:
                      type: integer
                      format: int32
                      minimum: 1
                      default: 1
                - name: limit
                  in: query
                  description: The number of cases to return
                  required: false
                  schema:
                      type: integer
                      format: int32
                      minimum: 1
                      maximum: 100
                      default: 10
                - name: count_limit
                  in: query
                  description: The maximum number of documents that will be counted in mongoDB to make queries faster
                  required: false
                  schema:
                      type: integer
                      format: int32
                      minimum: 100
                      default: 10000
                      maximum: 50000
                - name: sort_by
                  in: query
                  description: Keyword to sort by
                  required: false
                  schema:
                      type: string
                - name: order
                  in: query
                  description: Sorting order
                  required: false
                  schema:
                      type: string
                - name: q
                  in: query
                  description: The search query
                  required: false
                  allowEmptyValue: true
                  allowReserved: true
                  schema:
                      type: string
            responses:
                '200':
                    $ref: '#/components/responses/200CaseArray'
                '400':
                    $ref: '#/components/responses/400'
                '403':
                    $ref: '#/components/responses/403'
                '422':
                    $ref: '#/components/responses/422'
                '500':
                    $ref: '#/components/responses/500'
        post:
            tags: [Case]
            summary: Creates one (or multiple identical) new cases
            operationId: createCase
            requestBody:
                description: Case to add
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/NewCase'
            parameters:
                - name: validate_only
                  in: query
                  description: Whether to validate the case without creating it
                  required: false
                  schema:
                      type: boolean
                      default: false
                  allowEmptyValue: true
                - name: num_cases
                  in: query
                  description: The number of cases to create
                  required: false
                  schema:
                      type: integer
                      format: int32
                      minimum: 1
                      default: 1
            responses:
                '201':
                    $ref: '#/components/responses/201Create'
                '400':
                    $ref: '#/components/responses/400'
                '403':
                    $ref: '#/components/responses/403'
                '422':
                    $ref: '#/components/responses/422'
                '500':
                    $ref: '#/components/responses/500'
        put:
            tags: [Case]
            summary: Upserts a case
            description: >
                Creates or updates a case, depending on whether the provided values for
                the `caseReference.sourceId` and `caseReference.sourceEntryId` fields
                correspond to an existing case record.
            operationId: upsertCase
            requestBody:
                description: Case to upsert
                required: true
                content:
                    application/json:
                        schema:
                            allOf:
                                - $ref: '#/components/schemas/NewCase'
                            required:
                                - caseReference
            responses:
                '200':
                    $ref: '#/components/responses/200Case'
                '201':
                    $ref: '#/components/responses/201Case'
                '400':
                    $ref: '#/components/responses/400'
                '403':
                    $ref: '#/components/responses/403'
                '422':
                    $ref: '#/components/responses/422'
                '500':
                    $ref: '#/components/responses/500'
        delete:
            summary: >
                Deletes multiple cases. It is required to supply exactly one of either 
                caseIds or query in the request body. If caseIds are supplied, cases 
                corresponding to those caseIds will be deleted. If query is supplied, 
                all cases that match the query will be deleted.
            tags: [Case]
            operationId: deleteCases
            requestBody:
                description: Cases to delete
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                caseIds:
                                    description: Cases corresponding to these ids will be deleted
                                    type: array
                                    items:
                                        type: string
                                query:
                                    description: >
                                        Cases matching this query will be deleted. Must contain
                                        non-whitespace characters.
                                    type: string
                                    pattern: \S+
                            oneOf:
                                - required: [caseIds]
                                - required: [query]
            responses:
                '204':
                    description: Cases deleted
                '400':
                    $ref: '#/components/responses/400'
                '403':
                    $ref: '#/components/responses/403'
                '404':
                    $ref: '#/components/responses/404'
                '422':
                    $ref: '#/components/responses/422'
                '500':
                    $ref: '#/components/responses/500'
...
    /cases/downloadAsync:
        post:
            summary: >
                Streams an attachment of cases matching the search query.
            tags: [Case]
            operationId: downloadAsync
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                query:
                                    description: >
                                        Cases matching this query will be returned.
                                    type: string
            responses:
                '200':
                    $ref: '#/components/responses/200FilteredDataSetDownload'
                '204':
                    $ref: '#/components/responses/204'
                '500':
                    $ref: '#/components/responses/500'
...
    /cases/symptoms:
        get:
            summary: Lists most frequently used sypmtoms
            tags: [Suggest]
            operationId: listSymptoms
            parameters:
                - name: limit
                  in: query
                  description: The number of items to return
                  required: false
                  schema:
                      type: integer
                      format: int32
                      minimum: 1
                      maximum: 100
                      default: 5
            responses:
                '200':
                    $ref: '#/components/responses/200SymptomArray'
                '400':
                    $ref: '#/components/responses/400'
                '403':
                    $ref: '#/components/responses/403'
                '500':
                    $ref: '#/components/responses/500'
    /cases/placesOfTransmission:
        get:
            summary: Lists most frequently used places of transmission
            tags: [Suggest]
            operationId: listPlacesOfTransmission
            parameters:
                - name: limit
                  in: query
                  description: The number of items to return
                  required: false
                  schema:
                      type: integer
                      format: int32
                      minimum: 1
                      maximum: 100
                      default: 5
            responses:
                '200':
                    $ref: '#/components/responses/200PlacesOfTransmissionArray'
                '400':
                    $ref: '#/components/responses/400'
                '403':
                    $ref: '#/components/responses/403'
                '500':
                    $ref: '#/components/responses/500'
    /cases/occupations:
        get:
            summary: Lists most frequently used occupations
            tags: [Suggest]
            operationId: listOccupations
            parameters:
                - name: limit
                  in: query
                  description: The number of items to return
                  required: false
                  schema:
                      type: integer
                      format: int32
                      minimum: 1
                      maximum: 100
                      default: 5
            responses:
                '200':
                    $ref: '#/components/responses/200OccupationArray'
                '400':
                    $ref: '#/components/responses/400'
                '403':
                    $ref: '#/components/responses/403'
                '500':
                    $ref: '#/components/responses/500'
...
    /cases/batchUpsert:
        post:
            tags: [Case]
            summary: Upserts multiple cases
            operationId: batchUpsertCase
            requestBody:
                description: Cases to upsert
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CaseArray'
            responses:
                '200':
                    $ref: '#/components/responses/200BatchCaseUpsert'
                '207':
                    $ref: '#/components/responses/207BatchCaseUpsert'
                '500':
                    $ref: '#/components/responses/500'
...
components:
...
    responses:
...
        '200Case':
            description: OK
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Case'
        '200CaseArray':
            description: OK
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/CaseArray'
...
        '201Case':
            description: Created
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Case'
        '201Create':
            description: Created
            content:
                application/json:
                    schema:
                        anyOf:
                            - $ref: '#/components/schemas/Case'
                            - $ref: '#/components/schemas/CaseArray'
...
        '400':
            description: Malformed request
        '403':
            description: Forbidden
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
        '404':
            description: Not found
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
        '422':
            description: Unprocessable entity
        '500':
            description: Internal server error
servers:
    - url: http://localhost:3001/api
      description: Local server
    - url: http://dev-data.covid-19.global.health/api
      description: Development server
    - url: http://data.covid-19.global.health/api
      description: Production server